requirements.txt
python-dotenv
google-generativeai
torch
transformers
requests
customtkinter
